+++ CHANGES.rst
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+2.2.8 (unreleased)
+------------------
+
+Breaking changes:
+
+- *add item here*
+
+New features:
+
+- *add item here*
+
+Bug fixes:
+
+- *add item here*
+
+
 2.2.7 (2016-05-05)
 ------------------